[refactor][fix]: improve exception name static typings for errors sourced from async requests
Errors causes for operations backed by IDBRequests can now be discriminated against.

on the surface API. the errors will now show a generic type `boolean` parameter.

```
IDBIndexGetError<true> | IDBIndexGetError<false>
```
which represents whether the error was sourced from an IDBRequest or not, it is tied to the type of Error.`isFromRequest` property.
see `testObjectStoreErrorTypes` in `error-types.test.ts`